diesel_infer_query

Trait SchemaResolver

pub trait SchemaResolver {
    // Required methods
    fn resolve_field<'s>(
        &'s mut self,
        relation_schema: Option<&str>,
        query_relation: &str,
        field_name: &str,
    ) -> Result<&'s dyn SchemaField, Box<dyn Error + Send + Sync + 'static>>;
    fn list_fields<'s>(
        &'s mut self,
        relation_schema: Option<&str>,
        query_relation: &str,
    ) -> Result<Vec<&'s dyn SchemaField>, Box<dyn Error + Send + Sync + 'static>>;
}

A generic interface that allows this crate to request more information about certain database relations

Required Methods

fn resolve_field<'s>( &'s mut self, relation_schema: Option<&str>, query_relation: &str, field_name: &str, ) -> Result<&'s dyn SchemaField, Box<dyn Error + Send + Sync + 'static>>

Resolve a specific database field

fn list_fields<'s>( &'s mut self, relation_schema: Option<&str>, query_relation: &str, ) -> Result<Vec<&'s dyn SchemaField>, Box<dyn Error + Send + Sync + 'static>>

Get a list of fields for a specific database relation in the order returned by the database